extract manualTimer to tools/
